# public/swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "InsIIT API",
...
    "version": "1.0",
    "x-logo": {
      "url": "https://avatars.githubusercontent.com/u/146699003?s=1000&v=4"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": ""
    }
  ],
  "paths": {
    "/mess-menu": {
      "get": {
        "summary": "Get All Mess Menu Information",
        "tags": [
          "Mess Menu"
        ],
        "description": "Get Every information related to the Mess Menu",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Add Mess Menu",
        "tags": [
          "Mess Menu"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/mess-menu/{id}": {
      "summary": "Get Mess Menu by ID",
      "get": {
        "tags": [
          "Mess Menu"
        ],
        "description": "Get Mess Menu by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update Mess Menu by ID",
        "tags": [
          "Mess Menu"
        ],
        "description": "Update Mess Menu by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete Mess Menu by ID",
        "tags": [
          "Mess Menu"
        ],
        "description": "Delete Mess Menu by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/buses": {
      "summary": "Get All Buses Information",
      "get": {
        "tags": [
          "Buses"
        ],
        "description": "Get Every information related to the Buses",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Add Bus",
        "tags": [
          "Buses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/buses/{id}": {
      "summary": "Get Bus by ID",
      "get": {
        "summary": "Get Bus by ID",
        "tags": [
          "Buses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "summary": "Update Bus by ID",
        "tags": [
          "Buses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "BusName": {
                    "example": "any"
                  },
                  "DepartureTime": {
                    "example": "any"
                  },
                  "Destination": {
                    "example": "any"
                  },
                  "Source": {
                    "example": "any"
                  },
                  "Stops": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Bus by ID",
        "tags": [
          "Buses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/towns": {
      "get": {
        "summary": "Get All Towns/Stops",
        "tags": [
          "Buses"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/search": {
      "summary": "Search Buses by Source and Destination",
      "get": {
        "tags": [
          "Buses"
        ],
        "description": "",
        "parameters": [
          {
            "name": "source",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "destination",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events": {
      "summary": "Get All Events Information",
      "get": {
        "tags": [
          "Events"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Add Event",
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_name": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "start_time": {
                    "example": "any"
                  },
                  "poster_image_url": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "added_by": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/events/{id}": {
      "summary": "Get Event by ID",
      "get": {
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "summary": "Delete Event by ID",
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/events/{eventid}": {
      "put": {
        "summary": "Update Event by ID",
        "tags": [
          "Events"
        ],
        "description": "",
        "parameters": [
          {
            "name": "eventid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "event_name": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "date": {
                    "example": "any"
                  },
                  "start_time": {
                    "example": "any"
                  },
                  "poster_image_url": {
                    "example": "any"
                  },
                  "description": {
                    "example": "any"
                  },
                  "added_by": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/outlets": {
      "summary": "Get All Outlets Information",
      "get": {
        "tags": [
          "Outlets"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "summary": "Add Outlet",
        "tags": [
          "Outlets"
        ],
        "description": "",
        "parameters": [
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "landmark": {
                    "example": "any"
                  },
                  "open_time": {
                    "example": "any"
                  },
                  "close_time": {
                    "example": "any"
                  },
                  "rating": {
                    "example": "any"
                  },
                  "menu": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/outlets/{id}": {
      "summary": "Get Outlet by ID",
      "get": {
        "tags": [
          "Outlets"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "summary": "Update Outlet by ID",
        "tags": [
          "Outlets"
        ],
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "example": "any"
                  },
                  "location": {
                    "example": "any"
                  },
                  "landmark": {
                    "example": "any"
                  },
                  "open_time": {
                    "example": "any"
                  },
                  "close_time": {
                    "example": "any"
                  },
                  "rating": {
                    "example": "any"
                  },
                  "menu": {
                    "example": "any"
                  },
                  "image": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/outlets/menu/{outletId}": {
      "get": {
        "summary": "Get Outlet Menu by ID",
        "tags": [
          "Outlets"
        ],
        "description": "",
        "parameters": [
          {
            "name": "outletId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "Get the API Key from InsIIT Maintainers or mail us at metis@iitgn.ac.in"
      }
    }
  }
}